backend/requirements/dev.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
fastapi==0.104.1
fastapi-pagination==0.12.14
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
mariadb==1.1.8
packaging==23.2
pluggy==1.3.0
pydantic==2.5.2
pydantic_core==2.14.5
PyMySQL==1.1.0
pytest==7.4.4
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
